+++ assets/map/tilesets/Floor.tsx
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <tileset version="1.10" tiledversion="1.11.2" name="Floor" tilewidth="64" tileheight="64" tilecount="550" columns="22">
- <image source="../../graphics/tilemap/Floor.png" width="1408" height="1600"/>
+ <image source="../../graphics/tileimages/Floor.png" width="1408" height="1600"/>
  <wangsets>
   <wangset name="Water / Grass" type="corner" tile="-1">
    <wangcolor name="Water" color="#ff0000" tile="-1" probability="1"/>