+++ assets/map/tilesets/Grass.tsx
@@ -2,12 +2,12 @@
 <tileset version="1.10" tiledversion="1.11.2" name="grass" tilewidth="64" tileheight="64" tilecount="3" columns="0">
  <grid orientation="orthogonal" width="1" height="1"/>
  <tile id="0">
-  <image source="../../graphics/grass/grass_1.png" width="64" height="64"/>
+  <image source="../../graphics/tileimages/grass/grass_1.png" width="64" height="64"/>
  </tile>
  <tile id="1">
-  <image source="../../graphics/grass/grass_2.png" width="64" height="64"/>
+  <image source="../../graphics/tileimages/grass/grass_2.png" width="64" height="64"/>
  </tile>
  <tile id="2">
-  <image source="../../graphics/grass/grass_3.png" width="64" height="64"/>
+  <image source="../../graphics/tileimages/grass/grass_3.png" width="64" height="64"/>
  </tile>
 </tileset>