+++ code/enemy.py
@@ -1,16 +1,17 @@
 import pygame
 
-from data import monster_data
+from data.data import monster_data
 from entity import Entity
 from support import *
 
 class Enemy(Entity):
     def __init__(self, monster_name, pos, groups, obstacle_sprites,
                  damage_player, trigger_death_particles, add_exp):
-        super().__init__(groups)
-        self.sprite_type = 'enemy'
 
         self.import_graphics(monster_name)
+        super().__init__(groups, self.animations['idle'][0])
+        self.sprite_type = 'enemy'
+
         self.status = 'idle'
         self.image = self.animations[self.status][self.frame_index]
 
@@ -44,8 +45,8 @@ class Enemy(Entity):
         self.add_exp = add_exp
 
         # sounds
-        self.death_sound = pygame.mixer.Sound('../assets/audio/death.wav')
-        self.hit_sound = pygame.mixer.Sound('../assets/audio/hit.wav')
+        self.death_sound = pygame.mixer.Sound(get_assets_dir() + 'audio/death.wav')
+        self.hit_sound = pygame.mixer.Sound(get_assets_dir() + 'audio/hit.wav')
         self.attack_sound = pygame.mixer.Sound(monster_info['attack_sound'])
         self.death_sound.set_volume(0.6)
         self.hit_sound.set_volume(0.6)
@@ -54,7 +55,7 @@ class Enemy(Entity):
     # noinspection PyAttributeOutsideInit
     def import_graphics(self, name):
         self.animations = {'idle': [], 'move': [], 'attack': []}
-        main_path = f'../assets/graphics/monsters/{name}/'
+        main_path = get_assets_dir() + f'graphics/monsters/{name}/'
         for animation in self.animations.keys():
             self.animations[animation] = import_folder(main_path + animation)