+++ code/entity.py
@@ -2,9 +2,16 @@ from math import sin
 
 import pygame
 
+from settings import LAYERS
+
+
 class Entity(pygame.sprite.Sprite):
-    def __init__(self, groups: list[pygame.sprite.Group]):
+    def __init__(self, groups: list[pygame.sprite.Group], image):
         super().__init__(*groups)
+        self.image = image
+        self.rect = self.image.get_rect()  # Have to be replaced
+        self.hitbox = self.rect.copy()
+        self.z = LAYERS['main']
 
         # graphics
         self.frame_index = 0
@@ -12,17 +19,19 @@ class Entity(pygame.sprite.Sprite):
 
         # movement
         self.direction = pygame.math.Vector2()
+        self.obstacle_sprites = pygame.sprite.Group()  # Have to be replaced
 
     def move(self, speed):
         if self.direction.magnitude() != 0:
             self.direction = self.direction.normalize()
 
-        self.hitbox.x += self.direction.x * speed
+        self.rect.x += self.direction.x * speed
+        self.hitbox.centerx = self.rect.centerx
         self.collision('horizontal')
-        self.hitbox.y += self.direction.y * speed
-        self.collision('vertical')
 
-        self.rect.center = self.hitbox.center
+        self.rect.y += self.direction.y * speed
+        self.hitbox.centery = self.rect.centery
+        self.collision('vertical')
 
     def collision(self, direction):
         if direction == 'horizontal':
@@ -44,7 +53,4 @@ class Entity(pygame.sprite.Sprite):
     @staticmethod
     def wave_value():  # get alpha
         value = sin(pygame.time.get_ticks())
-        if value >= 0:
-            return 255
-        else:
-            return 0
+        return 255 if value >= 0 else 0