+++ code/magic.py
@@ -3,13 +3,14 @@ from random import randint
 import pygame
 
 from settings import *
+from support import get_assets_dir
 
 class MagicPlayer:
     def __init__(self, animation_player):
         self.animation_player = animation_player
         self.sounds = {
-            'heal': pygame.mixer.Sound('../assets/audio/heal.wav'),
-            'flame': pygame.mixer.Sound('../assets/audio/flame.wav')
+            'heal': pygame.mixer.Sound(get_assets_dir() + 'audio/heal.wav'),
+            'flame': pygame.mixer.Sound(get_assets_dir() + 'audio/flame.wav')
         }
 
     def heal(self, player, strength, cost, groups):