+++ code/settings.py
@@ -1,3 +1,5 @@
+from support import get_assets_dir
+
 # game setup
 WIDTH = 1280
 HEIGHT = 720
@@ -15,5 +17,13 @@ BAR_HEIGHT = 20
 HEALTH_BAR_WIDTH = 200
 ENERGY_BAR_WIDTH = 140
 ITEM_BOX_SIZE = 80
-UI_FONT = '../assets/fonts/joystix.ttf'
+UI_FONT = get_assets_dir() + 'fonts/joystix.ttf'
 UI_FONT_SIZE = 18
+
+LAYERS = {
+    'water': 0,
+    'floor': 1,
+    'main': 2,
+}
+
+del get_assets_dir