+++ code/tile.py
@@ -1,5 +1,6 @@
 import pygame
-from settings import TILE_SIZE, HITBOX_OFFSET
+
+from settings import TILE_SIZE, HITBOX_OFFSET, LAYERS
 
 class Tile(pygame.sprite.Sprite):
     def __init__(self, pos, groups: list[pygame.sprite.Group], sprite_type,
@@ -8,6 +9,7 @@ class Tile(pygame.sprite.Sprite):
         self.image = surface
         self.sprite_type = sprite_type
         y_offset = HITBOX_OFFSET[sprite_type]
+        self.z = LAYERS['floor']
 
         if sprite_type == 'object':
             self.rect = self.image.get_rect(topleft=(pos[0], pos[1] - TILE_SIZE))