+++ docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+- Restructured the project
+- Added a `docs` folder
+- Migrated from `pygame` to `pygame-ce`
+- Use TMX files for maps instead of CSV files (added `PyTMX` dependency)
+- Removed the `test` folder from `graphics`
+- Got rid of the logic to draw `ground.png` and deleted that image
+- Added layers logic to draw the objects correctly
+- Write docs
+
 ## [0.1.0] - 2025-12-2
 - Initial release